+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog(https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning(https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Support WeasyPrint options (`--info`, `--quiet`, `--verbose`, `--debug`, `--version`, `--no-http-redirects`, `--fail-on-http-errors`)
 
 ## 2.4.0 - 2026-01-20
 ### Added